+++ src/links.rs
@@ -39,4 +39,5 @@ pub mod statuses {
     pub const LOOKUP: &'static str = "https://api.twitter.com/1.1/statuses/lookup.json";
     pub const HOME_TIMELINE: &'static str = "https://api.twitter.com/1.1/statuses/home_timeline.json";
     pub const MENTIONS_TIMELINE: &'static str = "https://api.twitter.com/1.1/statuses/mentions_timeline.json";
+    pub const USER_TIMELINE: &'static str = "https://api.twitter.com/1.1/statuses/user_timeline.json";
 }